+++ src/lib/libast/sfio/_sfputm.c
@@ -47,7 +47,7 @@ ssize_t _sfputm(Sfio_t* f,	/* write a portable ulong to this stream */
 	n = (ps-s)+1;
 
 	if(n > 8 || SFWPEEK(f,ps,p) < n)
-		n = SFWRITE(f,s,n); /* write the hard way */
+		n = SFWRITE(f,s,(size_t)n); /* write the hard way */
 	else
 	{	switch(n)
 		{
