+++ src/lib/libast/stdio/fread.c
@@ -24,5 +24,5 @@ fread(void* p, size_t s, size_t n, Sfio_t* f)
 {
 	ssize_t	v;
 
-	return ((v = sfread(f, p, s * n)) <= 0) ? 0 : (v / s);
+	return ((v = sfread(f, p, s * n)) <= 0) ? 0 : ((size_t)v / s);
 }
