+++ src/lib/libsum/sum-prng.c
@@ -68,11 +68,11 @@ prng_open(const Method_t* method, const char* name)
 			if (*s == '=' && !v)
 				v = s;
 		i = (v ? v : s) - t;
-		if (isdigit(*t) || v && strneq(t, "mpy", i) && (t = v + 1))
+		if (isdigit(*t) || v && strneq(t, "mpy", (size_t)i) && (t = v + 1))
 			sum->mpy = (Prngnum_t)strtoul(t, NULL, 0);
-		else if (strneq(t, "add", i))
+		else if (strneq(t, "add", (size_t)i))
 			sum->add = v ? (Prngnum_t)strtoul(v + 1, NULL, 0) : ~sum->add;
-		else if (strneq(t, "init", i))
+		else if (strneq(t, "init", (size_t)i))
 			sum->init = v ? (Prngnum_t)strtoul(v + 1, NULL, 0) : ~sum->init;
 		if (*s == '-')
 			s++;
